feat(SPV-1398): move transactions endpoints to v2
