+++ actions/v2/register.go
@@ -3,6 +3,7 @@ package v2
 import (
 	"github.com/bitcoin-sv/spv-wallet/actions/v2/admin"
 	"github.com/bitcoin-sv/spv-wallet/actions/v2/operations"
+	"github.com/bitcoin-sv/spv-wallet/actions/v2/transactions"
 	"github.com/bitcoin-sv/spv-wallet/actions/v2/users"
 	"github.com/bitcoin-sv/spv-wallet/server/handlers"
 )
@@ -11,6 +12,7 @@ import (
 func Register(handlersManager *handlers.Manager) {
 	users.RegisterRoutes(handlersManager)
 	operations.RegisterRoutes(handlersManager)
+	transactions.RegisterRoutes(handlersManager)
 
 	admin.Register(handlersManager)
 }
