+++ actions/v2/transactions/outlines_record.go
@@ -1,7 +1,7 @@
 package transactions
 
 import (
-	"github.com/bitcoin-sv/spv-wallet/actions/transactions/internal/mapping/annotatedtx"
+	"github.com/bitcoin-sv/spv-wallet/actions/v2/transactions/internal/mapping/annotatedtx"
 	"github.com/bitcoin-sv/spv-wallet/engine/spverrors"
 	"github.com/bitcoin-sv/spv-wallet/server/reqctx"
 	"github.com/gin-gonic/gin"
